fix(deps): update all non-major dependencies (minor) (#25)
* fix(deps): update all non-major dependencies

* fix: adapt to esp-hub75 0.13 and sntpc 0.11 breaking changes

- Hub75::render now takes &impl FrameBuffer; pass a shared reference
- sntpc NtpResult::seconds is now u64; narrow the derived offset to u32


* refactor: drop Option wrapper around hub75 frame buffer

esp-hub75 0.13's render() takes a shared reference, so the buffer no
longer needs Option::replace to swap. Matches the upstream lcd_cam_bp
example and removes the &**fb deref.


---------

+++ src/clock.rs
@@ -114,7 +114,8 @@ impl<'a, I2C: I2c> Clock<'a, I2C> {
         if let Ok(response) = sntp_process_response(addr, socket_ref, context, req).await {
             println!("received NTP response: {:?}", response);
             let uptime_seconds = Instant::now().as_secs() as u32;
-            let boot_time_offset = response.seconds.saturating_sub(uptime_seconds);
+            // sntpc >= 0.11 reports seconds as u64; the offset fits in u32 well past 2100.
+            let boot_time_offset = response.seconds.saturating_sub(uptime_seconds as u64) as u32;
             TIME_OFFSET_SECONDS.store(boot_time_offset, Ordering::Relaxed);
             self.set_rtc();
         } else {
